+++ R/write-beast.R
@@ -116,10 +116,159 @@ write_beast_newick <- function(treedata, file = "",
     cat(res, file = file, append = append, sep = "\n")
 }
 
+.write.tree3 <- function(phy, digits = 10, tree.prefix = "", node_anno = NULL) {
+    node.label <- checkLabel(phy$node.label)
+    if (length(node.label) == 0) node.label <- NULL
+
+    .write.tree4(phy$edge, digits = digits, tree.prefix = tree.prefix,
+                 root.edge = phy$root.edge, edge.length = phy$edge.length,
+                 tip.label = checkLabel(phy$tip.label),
+                 node.label = node.label,
+                 node_anno = node_anno)
+}
+
+## derived from .write.tree3_old to tailor for edge list, then all tree-like graph can be supported.
+.write.tree4 <- function(edge, digits = 10, tree.prefix = "", root.edge=NULL, edge.length = NULL,
+                         tip.label=NULL, node.label=NULL, id_as_label = FALSE, node_anno = NULL) {
+
+    edge.label <- edge
+    edge <- matrix(as.numeric(as.factor(edge.label)), ncol=2)
+
+    f.d <- paste("%.", digits, "g", sep = "")
+
+    to_tiplab <- function(edge, i) {
+        edge.label[edge[,2] == i, 2][1]
+    }
+
+    to_nodelab <- function(edge, i) {
+        edge.label[edge[,1] == i, 1][1]
+    }
+
+    brl <- !is.null(edge.length)
+    nodelab <- !is.null(node.label)
+
+    if (id_as_label) nodelab <- TRUE
+
+    cp <- function(x){
+        STRING[k] <<- x
+        k <<- k + 1
+    }
+
+    add.internal <- function(i) {
+        cp("(")
+        desc <- kids[[i]]
+        for (j in desc) {
+            if (j %in% edge[,1]) add.internal(j)
+            ## if (j > n) add.internal(j)
+            else add.terminal(ind[j])
+            if (j != desc[length(desc)]) cp(",")
+        }
+        cp(")")
+        ## if (nodelab && i > n) {
+        if (nodelab && i %in% edge[,1]) {
+            ## cp(phy$node.label[i - n]) # fixed by Naim Matasci (2010-12-07)
+            if (is.null(node_anno)) {
+                if (id_as_label) nl <- to_nodelab(edge, i)
+                else nl <- node.label[i - n]
+            } else if (!is.na(node_anno[i])) {
+                if (id_as_label) nl <- paste0(to_nodelab(edge, i), node_anno[i])
+                else nl <- paste0(node.label[i - n], node_anno[i])
+            }
+            cp(nl)
+        } else if (i %in% edge[,1] && !is.null(node_anno)) {
+            cp(node_anno[i])
+        }
+        if (brl) {
+            cp(":")
+            cp(sprintf(f.d, edge.length[ind[i]]))
+        }
+    }
+
+    add.terminal <- function(i) {
+        ii <- edge[i, 2]
+        if (is.null(node_anno) || is.na(node_anno[ii])) {
+            if (id_as_label) tl <- to_tiplab(edge, ii)
+            else tl <- tip.label[ii]
+        } else {
+            if (id_as_label) tl <- paste0(to_tiplab(edge, ii), node_anno[ii])
+            else tl <- paste0(tip.label[ii], node_anno[ii])
+        }
+        cp(tl)
+        if (brl) {
+            cp(":")
+            cp(sprintf(f.d, edge.length[i]))
+        }
+    }
+
+    Ntip <- function(edge) {
+        tip <- edge[,2][!edge[,2] %in% edge[,1]]
+        length(tip)
+    }
+
+    Nnode <- function(edge) {
+        tip <- edge[,2][!edge[,2] %in% edge[,1]]
+        node <- edge[,1][!edge[,1] %in% tip]
+        length(node)
+    }
+
+    ## borrowed from phangorn:
+    parent <- edge[, 1]
+    children <- edge[, 2]
+    n <- Ntip(edge)
+    kids <- vector("list", n + Nnode(edge))
+
+    for (i in seq_along(parent))
+        kids[[parent[i]]] <- c(kids[[parent[i]]], children[i])
+
+    ind <- match(1:max(edge), edge[, 2])
+
+    LS <- 4*n + 5
+    if (brl) LS <- LS + 4*n
+    if (nodelab)  LS <- LS + n
+    STRING <- character(LS)
+    k <- 1
+    cp(tree.prefix)
+    cp("(")
+    getRoot <- function(edge)
+        edge[, 1][!match(edge[, 1], edge[, 2], 0)][1]
+    root <- getRoot(edge) # replaced n+1 with root - root has not be n+1
+    desc <- kids[[root]]
+    for (j in desc) {
+        if (j %in% edge[,1]) add.internal(j)
+        ## if (j > n) add.internal(j)
+        else add.terminal(ind[j])
+        if (j != desc[length(desc)]) cp(",")
+    }
+
+    if (is.null(root.edge)) {
+        cp(")")
+        if (nodelab) {
+            if (!is.null(node_anno) && !is.na(node_anno[root])) {
+                if(id_as_label) cp(paste0(to_nodelab(edge, root), node_anno[root]))
+                else cp(paste0(node.label[1], node_anno[root]))
+            } else {
+                if (id_as_label) cp(to_nodelab(edge, root))
+                else cp(node.label[1])
+            }
+        } else if (!is.null(node_anno) && !is.na(node_anno[root])) {
+            cp(node_anno[root])
+        }
+        cp(";")
+    } else {
+        cp(")")
+        if (nodelab) cp(node.label[1])
+        cp(":")
+        cp(sprintf(f.d, root.edge))
+        cp(";")
+    }
+    paste(STRING, collapse = "")
+}
+
+
 ## this function was derived from ape:::.write.tree2
 ## by adding node/tip label with associated annotation in BEAST format
 ##' @importFrom ape checkLabel
-.write.tree3 <- function(phy, digits = 10, tree.prefix = "", node_anno = NULL) {
+.write.tree3_old <- function(phy, digits = 10, tree.prefix = "", node_anno = NULL) {
     brl <- !is.null(phy$edge.length)
     nodelab <- !is.null(phy$node.label)
     phy$tip.label <- checkLabel(phy$tip.label)